+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log
 
+### 0.0.2 (2022-12-15)
+
+#### Features
+
+- Add reporting categories.
+- Delete script files and put all scripts in the popup.js file.
+- Add the function of like clicking.
+
 ### 0.0.1 (2022-11-26)
 
 #### Features